+++ common/v2/features/Home/Home.tsx
@@ -12,7 +12,7 @@ import {
   BottomActionPanel,
   FeaturesPanel
 } from './components';
-import { getFeaturedOS } from 'v2/features/helpers';
+import { getFeaturedOS } from 'v2/utils';
 import { GithubService } from 'v2/services';
 import { COLORS, BREAK_POINTS } from 'v2/theme';
 import { GITHUB_RELEASE_NOTES_URL as DEFAULT_LINK } from 'v2/config';
