+++ common/v2/features/index.ts
@@ -1,7 +1,5 @@
 export * from './Wallets';
 export { ScreenLockLocking } from './ScreenLock';
-export { fetchGasPriceEstimates, getDefaultEstimates } from './Gas';
-export { getNonce } from './Nonce';
 export { Drawer, AccountScreen, UnlockScreen, SelectLanguage } from './Drawer';
 export { DevTools } from './DevTools';
 export { Home } from './Home';
