+++ common/v2/services/Nonce/nonceFunctions.ts
@@ -1,5 +1,5 @@
-import { ExtendedAccount } from 'v2/services/Account/types';
-import { Network } from 'v2/services/Network/types';
+import { ExtendedAccount } from '../Account/types';
+import { Network } from '../Network/types';
 import ProviderHandler from 'v2/config/networks/providerHandler';
 
 export function getNonce(network: Network, account: ExtendedAccount) {
