+++ common/v2/utils/index.ts
@@ -2,4 +2,5 @@ export * from './environment';
 export * from './error';
 export * from './platform';
 export * from './validators';
+export { getFeaturedOS } from './getFeaturedOS';
 export { default as getSymbolIcon } from './symbolIcon';
